Issue where with stmt raises when exc_info is set (#345)
Found this issue while working on something else: if exc_info is set
(e.g. in an except clause) then the finally part of the with statement
will re-raise that exception because it thinks the with body was
responsible for the exception.

Fix is to check whether πE is set instead of checking exc_info when
entering the finally part of the with block.